fix: 불필요한 format 추상화 제거
